+++ MCSC/ViewModels/Routing/ActionRegistry.swift
@@ -19,4 +19,5 @@ final class ActionRegistry {
     let almostMaximizeAction = AlmostMaximizeAction()
     let closeAllTabsAction = CloseAllTabsAction()
     let newWindowAction = NewWindowAction()
+    let ejectVolumeAction = EjectVolumeAction()
 }